First Commit Of Customized Pong
I switched the main.py and pong.kv files around and came up with a new
Python file randomLaunchDeg.py.

It prints random numbers but the numbers are appropriate for the pong
ball to launch in the main.py file. Check it out!

Previously: The ball would launch in any direction, including up and
down so it would've taken forver.

Now: The ball only launches towards the left and right walls to make the
game more fun.

Next Up: I'm going to add color effects every time the player scores a
point because it is almost impossible to win this game.